+++ ulib/src/main/java/player/efis/common/AirspaceClass.java
@@ -3,7 +3,7 @@ package player.efis.common;
 import java.util.*;
 
 
-enum AirspaceClass
+enum AirspaceClass_t
 {
     CLASSA,
     CLASSB,
@@ -23,4 +23,26 @@ enum AirspaceClass
     MATZ,
     GWAVE,
     OTHER
+}
+
+public class AirspaceClass
+{
+    public static boolean A; //CLASSA;
+    public static boolean B; //CLASSB;
+    public static boolean C; //CLASSC;
+    public static boolean D; //CLASSD;
+    public static boolean E; //CLASSE;
+    public static boolean F; //CLASSF;
+    public static boolean G; //CLASSG;
+    public static boolean R; //RESTRICT;
+    public static boolean Q; //DANGER;
+    public static boolean P;// PROHIBITED;
+    public static boolean WAVE;
+    public static boolean NOGLIDER;
+    public static boolean CTR;
+    public static boolean TMZ;
+    public static boolean RMZ;
+    public static boolean MATZ;
+    public static boolean GWAVE;
+    public static boolean OTHER;
 }